heartbeat stopping

+++ quadrupel-dronecode/src/main.rs
@@ -111,7 +111,7 @@ fn main() -> ! {
         }
 
         //Check heartbeat
-        if (Motors::get_time_us() - last_message_time) > (HEARTBEAT_FREQ * HEARTBEAT_TIMEOUT_MULTIPLE) {
+        if (Motors::get_time_us() - last_message_time) > (HEARTBEAT_FREQ * HEARTBEAT_TIMEOUT_MULTIPLE) as u32 {
             state.set_mode(Mode::Panic);
         }
